backend/requirements.txt
# ========================================
# HECTIC SWARM Dependencies
# ========================================

# Web Framework
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.0
python-dotenv==1.0.0

# AI/LLM
openai==1.40.0
httpx==0.27.0
tenacity==8.5.0
requests==2.31.0  # For MCP tool HTTP calls

# Database & RAG
psycopg[binary]==3.2.1  # PostgreSQL driver for Python 3.13
pgvector==0.3.0         # Vector similarity search
langchain==0.1.16       # RAG orchestration
langchain-openai==0.1.1 # OpenAI embeddings
tiktoken==0.6.0         # Token counting

# Swarm Coordination (optional - uncomment if using swarms library)
# swarms==5.0.0

# Hive-Mind Swarm Database
# SQLite is built-in to Python (no dependency needed)
# Pydantic 2.8.0 already included above for schema validation

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0

# Code Quality
black==24.3.0
mypy==1.9.0
ruff==0.3.5             # Fast Python linter
types-psycopg2==2.9.21  # Type stubs for mypy

# Subprocess & Build Integration
subprocess32==3.5.4; python_version < '3.0'  # Backport for older Python
